+++ ql/src/java/org/apache/hadoop/hive/ql/ddl/table/constraint/add/AlterTableAddConstraintAnalyzer.java
@@ -34,7 +34,11 @@ import org.apache.hadoop.hive.ql.ddl.DDLSemanticAnalyzerFactory.DDLType;
 import org.apache.hadoop.hive.ql.ddl.table.AbstractAlterTableAnalyzer;
 import org.apache.hadoop.hive.ql.ddl.table.constraint.Constraints;
 import org.apache.hadoop.hive.ql.ddl.table.constraint.ConstraintsUtils;
+import org.apache.hadoop.hive.ql.ddl.table.partition.add.AlterTableAddPartitionDesc;
+import org.apache.hadoop.hive.ql.exec.Task;
 import org.apache.hadoop.hive.ql.exec.TaskFactory;
+import org.apache.hadoop.hive.ql.io.AcidUtils;
+import org.apache.hadoop.hive.ql.metadata.Table;
 import org.apache.hadoop.hive.ql.parse.ASTNode;
 import org.apache.hadoop.hive.ql.parse.HiveParser;
 import org.apache.hadoop.hive.ql.parse.SemanticException;
@@ -43,43 +47,19 @@ import org.apache.hadoop.hive.ql.parse.SemanticException;
  * Analyzer for add constraint commands.
  */
 @DDLType(types = HiveParser.TOK_ALTERTABLE_ADDCONSTRAINT)
-public class AlterTableAddConstraintAnalyzer extends AbstractAlterTableAnalyzer {
+public class AlterTableAddConstraintAnalyzer extends AbstractAddConstraintAnalyzer {
   public AlterTableAddConstraintAnalyzer(QueryState queryState) throws SemanticException {
     super(queryState);
   }
 
   @Override
-  protected void analyzeCommand(TableName tableName, Map<String, String> partitionSpec, ASTNode command)
+  protected void postProcess(Table table, AlterTableAddConstraintDesc desc)
       throws SemanticException {
-    // TODO CAT - for now always use the default catalog.  Eventually will want to see if
-    // the user specified a catalog
-    List<SQLPrimaryKey> primaryKeys = new ArrayList<>();
-    List<SQLForeignKey> foreignKeys = new ArrayList<>();
-    List<SQLUniqueConstraint> uniqueConstraints = new ArrayList<>();
-    List<SQLCheckConstraint> checkConstraints = new ArrayList<>();
-
-    ASTNode constraintNode = (ASTNode)command.getChild(0);
-    switch (constraintNode.getToken().getType()) {
-    case HiveParser.TOK_UNIQUE:
-      ConstraintsUtils.processUniqueConstraints(tableName, constraintNode, uniqueConstraints);
-      break;
-    case HiveParser.TOK_PRIMARY_KEY:
-      ConstraintsUtils.processPrimaryKeys(tableName, constraintNode, primaryKeys);
-      break;
-    case HiveParser.TOK_FOREIGN_KEY:
-      ConstraintsUtils.processForeignKeys(tableName, constraintNode, foreignKeys);
-      break;
-    case HiveParser.TOK_CHECK_CONSTRAINT:
-      ConstraintsUtils.processCheckConstraints(tableName, constraintNode, null, checkConstraints, command,
-          ctx.getTokenRewriteStream());
-      break;
-    default:
-      throw new SemanticException(ErrorMsg.NOT_RECOGNIZED_CONSTRAINT.getMsg(constraintNode.getToken().getText()));
+    if (!AcidUtils.isTransactionalTable(table)) {
+      return;
     }
 
-    Constraints constraints = new Constraints(primaryKeys, foreignKeys, null, uniqueConstraints, null,
-        checkConstraints);
-    AlterTableAddConstraintDesc desc = new AlterTableAddConstraintDesc(tableName, null, constraints);
-    rootTasks.add(TaskFactory.get(new DDLWork(getInputs(), getOutputs(), desc)));
+    setAcidDdlDesc(desc);
   }
+
 }
